Move file saving to local dev script
